chore/rename file for consistency
